+++ database/migrations/2017_04_08_163453_create_world_cities_table.php
@@ -14,9 +14,9 @@ class CreateWorldCitiesTable extends Migration
     public function up()
     {
         Schema::create('world_cities', function (Blueprint $table) {
-            $table->increments('id')->comment('Auto increase ID');
-            $table->integer('country_id')->unsigned()->comment('Country ID');
-            $table->integer('division_id')->unsigned()->nullable()->index('division_id')->comment('Division ID');
+            $table->bigIncrements('id')->comment('Auto increase ID');
+            $table->bigInteger('country_id')->unsigned()->comment('Country ID');
+            $table->bigInteger('division_id')->unsigned()->nullable()->index('division_id')->comment('Division ID');
             $table->string('name', 255)->default('')->comment('City Name');
             $table->string('full_name', 255)->nullable()->comment('City Fullname');
             $table->string('code', 64)->nullable()->comment('City Code');
